SpaceX - Data Collection API
The goal of this project is to predict if the Falcon 9 First Stage will successfully land. The selling point for SpaceX is that they have made it so they can reuse their first stage and lower the cost per launch. SpaceX lists a cost of 62 million dollars, while competitors launches cost nearly 165 million dollars. SpaceX has launched many rockets, and not all of them will be able to be reused. By determining if the SpaceX rocket will land, we can better understand the cost of a SpaceX launch vs their competitors. In this first notebook, we will use the SpaceX API to collect the data and we will do some minor wrangling to get it in the format we will need to work with.